+++ ansible_collections/scruffaluff/bootware/roles/caffeine/files/caffeinate.ps1
@@ -5,6 +5,8 @@
 
 # Exit immediately if a PowerShell cmdlet encounters an error.
 $ErrorActionPreference = 'Stop'
+# Disable progress bar for PowerShell cmdlets.
+$ProgressPreference = 'SilentlyContinue'
 # Exit immediately when an native executable encounters an error.
 $PSNativeCommandUseErrorActionPreference = $True